+++ Economy/BudgetWinForms/Forms/PurchaseForm.cs
@@ -14,12 +14,17 @@ namespace BudgetWinForms
     {
         public int IdPurchase { get; private set; }
 
+        #region Constructors
+
         public PurchaseForm()
         {
             InitializeComponent();
             UpdateComboBox();
             UpdateDataGrid();
         }
+
+        #endregion
+        #region Metods
 
         private void UpdateComboBox()
         {
@@ -31,6 +36,7 @@ namespace BudgetWinForms
                     var item = new ListBoxItem(source.Id, source.Name);
                     purchaseComboBox.Items.Add(item);
                 }
+                purchaseComboBox.Text = "Выберите источник";
             }
         }
         
@@ -43,30 +49,51 @@ namespace BudgetWinForms
                 var renameButton = new DataGridViewButtonCell();
                 var removeButton = new DataGridViewButtonCell();
                 foreach (var purchase in purchases)
-                    purchaseDataGridView.Rows.Add(purchase.Id,purchase.DateTime, purchase.Source.Name, renameButton, removeButton);
+                    purchaseDataGridView.Rows.Add(
+                        purchase.Id,
+                        purchase.DateTime, 
+                        purchase.Source.Name, 
+                        renameButton, 
+                        removeButton
+                    );
             }
         }
 
-        private void purchaseButton_Click(object sender, EventArgs e)
+        #endregion
+        #region Events Handlers
+
+        private void purchaseButtonClick(object sender, EventArgs e)
         {
             using (var db = new BudgetModel())
             {
                 var purchase = new Purchase();
                 purchase.DateTime = purchaseDateTimePicker.Value;
                 var selectedSource = purchaseComboBox.SelectedItem as ListBoxItem;
+                if (selectedSource == null)
+                {
+                    MessageBox.Show("Выберите источник");
+                    return;
+                }
                 purchase.Source = db.Sources.Find(selectedSource.Id);
                 db.Purchases.Add(purchase);
                 db.SaveChanges();
                 IdPurchase = purchase.Id;
                 ChekItemForm chekItemForm = new ChekItemForm(IdPurchase);
-                chekItemForm.ShowDialog(this);
-                UpdateDataGrid();
+                if (chekItemForm.ShowDialog(this) == DialogResult.OK)
+                    UpdateDataGrid();
+                else
+                {
+                    db.Purchases.Remove(purchase);
+                    db.SaveChanges();
+                    UpdateDataGrid();
+                }
             }
         }
 
-        private void purchaseDataGridView_CellContentClick(object sender, DataGridViewCellEventArgs e)
+        private void purchaseDataGridViewCellContentClick(object sender, DataGridViewCellEventArgs e)
         {
-            if(e.ColumnIndex == 3)
+            if (e.RowIndex == -1) return;
+            if (e.ColumnIndex == 3)
             {
                 IdPurchase = (int)(purchaseDataGridView[0, e.RowIndex].Value);
                 ChekItemForm chekItemForm = new ChekItemForm(IdPurchase);
@@ -74,18 +101,25 @@ namespace BudgetWinForms
             }
             else if(e.ColumnIndex == 4)
             {
-                IdPurchase = (int)(purchaseDataGridView[0, e.RowIndex].Value);
-                using (var db = new BudgetModel())
+                var res = MessageBox.Show("Удалить?", "Удаление элемента", MessageBoxButtons.YesNo);
+                if (res == DialogResult.No) return;
+                else
                 {
-                    var removePurchase = db.Purchases.Find(IdPurchase);
-                    var removeChekItems = db.Purchases.Find(IdPurchase).ChekItems.ToList();
-                    foreach (var removeChekItem in removeChekItems)
-                        db.ChekItems.Remove(removeChekItem);
-                    db.Purchases.Remove(removePurchase);
-                    db.SaveChanges();
+                    IdPurchase = (int)(purchaseDataGridView[0, e.RowIndex].Value);
+                    using (var db = new BudgetModel())
+                    {
+                        var removePurchase = db.Purchases.Find(IdPurchase);
+                        var removeChekItems = db.Purchases.Find(IdPurchase).ChekItems.ToList();
+                        foreach (var removeChekItem in removeChekItems)
+                            db.ChekItems.Remove(removeChekItem);
+                        db.Purchases.Remove(removePurchase);
+                        db.SaveChanges();
+                    }
                 }
             }
             UpdateDataGrid();
         }
+
+        #endregion
     }
 }
